+++ noodles-fasta/examples/fasta_write.rs
@@ -8,7 +8,7 @@ use noodles_fasta as fasta;
 
 fn main() -> io::Result<()> {
     let stdout = io::stdout().lock();
-    let mut writer = fasta::Writer::new(stdout);
+    let mut writer = fasta::io::Writer::new(stdout);
 
     let definition = fasta::record::Definition::new("sq0", None);
     let sequence = fasta::record::Sequence::from(b"ACGT".repeat(64));